requirements.txt
streamlit
numpy==2.3.4
pandas
scipy
scikit-learn>=1.4.2
joblib
pillow
easyocr
opencv-python-headless


